server: Change some score variables to type double. Points still have integer values.

+++ src/server/score.h
@@ -33,7 +33,7 @@
 
 /* score.c */
 
-void Score(player_t *pl, int points, clpos_t pos, const char *msg);
+void Score(player_t *pl, double points, clpos_t pos, const char *msg);
 int Rate(int winner, int loser);
 
 /*
@@ -51,8 +51,9 @@ int Rate(int winner, int loser);
  * KK 28-4-98: Same for killing your own tank.
  * KK 7-11-1: And for killing a member of your alliance
  */
-void Score_players(player_t *winner_pl, int winner_score, char *winner_msg,
-                   player_t *loser_pl, int loser_score, char *loser_msg);
+void Score_players(player_t *winner_pl, double winner_score,
+                   char *winner_msg, player_t *loser_pl,
+                   double loser_score, char *loser_msg);
 
 int Get_Score(player_t *pl);
